# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZenMon API Documentation",
        "description": "Monitoring Application - Alternative to Zabbix/Nagios",
        "contact": {
            "email": "admin@zenmon.local"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001",
            "description": "ZenMon API Server"
        }
    ],
    "paths": {
...
        "/api/metric-types": {
            "get": {
                "tags": [
                    "Metric Types"
                ],
                "summary": "Get list of metric types",
                "description": "Returns list of available metric types with optional filtering",
                "operationId": "getMetricTypesList",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by metric name or description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "unit",
                        "in": "query",
                        "description": "Filter by unit",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MetricType"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Metric Types"
                ],
                "summary": "Create new metric type",
                "description": "Create new metric type definition",
                "operationId": "storeMetricType",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "metric_name",
                                    "unit"
                                ],
                                "properties": {
                                    "metric_name": {
                                        "description": "Unique metric name",
                                        "type": "string",
                                        "example": "CPU"
                                    },
                                    "unit": {
                                        "description": "Unit of measurement",
                                        "type": "string",
                                        "example": "%"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "CPU usage percentage"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Metric type created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/MetricType"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/metric-types/stats": {
            "get": {
                "tags": [
                    "Metric Types"
                ],
                "summary": "Get metric types with statistics",
                "description": "Returns metric types with usage statistics for dashboard",
                "operationId": "getWithStats",
                "responses": {
                    "200": {
                        "description": "Metric types with statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "allOf": [
                                                    {
                                                        "$ref": "#/components/schemas/MetricType"
                                                    },
                                                    {
                                                        "properties": {
                                                            "metrics_count": {
                                                                "type": "integer"
                                                            },
                                                            "active_hosts_count": {
                                                                "type": "integer"
                                                            },
                                                            "last_metric_date": {
                                                                "type": "string",
                                                                "format": "date-time"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/metric-types/units": {
            "get": {
                "tags": [
                    "Metric Types"
                ],
                "summary": "Get available units",
                "description": "Returns list of commonly used units for creating metric types",
                "operationId": "getAvailableUnits",
                "responses": {
                    "200": {
                        "description": "Available units",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "unit": {
                                                        "type": "string",
                                                        "example": "%"
                                                    },
                                                    "description": {
                                                        "type": "string",
                                                        "example": "Percentage"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": "Usage"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MetricType": {
                "title": "MetricType",
                "description": "Type of metric (CPU, RAM, Disk, Network)",
                "properties": {
                    "metric_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "metric_name": {
                        "type": "string",
                        "example": "CPU"
                    },
                    "unit": {
                        "type": "string",
                        "example": "%"
                    },
                    "description": {
                        "type": "string",
                        "example": "CPU usage percentage"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}